Make sure we don't delete groups whose jobs are too young

+++ async/api.py
@@ -87,8 +87,10 @@ def remove_old_jobs(remove_jobs_before_days=30, resched_hours=8):
         Q(executed__lt=start_remove_jobs_before_dt)).delete()
     groups = Group.objects.filter(Q(jobs__executed__isnull=False))
     for group in groups.iterator():
-        group.jobs.all().delete()
-        group.delete()
+        group.jobs.filter(Q(executed__isnull=False),
+        Q(executed__lt=start_remove_jobs_before_dt)).delete()
+        if group.jobs.count() == 0:
+            group.delete()
 
     next_exec = _get_today_dt() + timedelta(hours=resched_hours)
